chore: move things to individual files
